+++ googleform-to-backlog/CHANGES.md
@@ -2,4 +2,6 @@
 
 ## Next
 
+- Feature: Created sessions to backlog script
+- Refactor: Renamed main.js to sponser.js
 - Feature: Created initial script